+++ HelperFunctions.py
